pyproject.toml
[project]
name = "tesla-crash-synth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "diffusers>=0.36.0",
    "gradio>=6.5.1",
    "numpy>=2.4.2",
    "openai>=2.18.0",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.1.0",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "jupyter>=1.1.1",
    "pytest>=9.0.2",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["config*", "utils*"]
